completed dto statistics testing with postman

+++ server/allTheMethods/src/main/java/com/example/allTheMethods/dto/response/DifficultyStatsResponse.java
@@ -2,6 +2,6 @@ package com.example.allTheMethods.dto.response;
 
 public record DifficultyStatsResponse(
         String difficulty,
-        int solvedCount
+        Long solvedCount
 ) {
 }
